Refactoring and new seekers

+++ src/main/kotlin/de/goddchen/hiddennumbers/seekers/DayOfYearSeeker.kt
@@ -6,8 +6,8 @@ import java.time.LocalDate
 
 @Component
 @Qualifier("LocalDateSeeker")
-class DayOfYearSeeker(@Qualifier("LongSeeker") private val longSeekers: List<Seeker<Int>>) : Seeker<LocalDate> {
+class DayOfYearSeeker(@Qualifier("LongSeeker") private val longSeekers: List<Seeker<Long>>) : Seeker<LocalDate> {
     override fun findMagic(data: LocalDate) =
-        longSeekers.flatMap { it.findMagic(data.dayOfYear) }
+        longSeekers.flatMap { it.findMagic(data.dayOfYear.toLong()) }
             .map { it.apply { args["dayOfYear"] = data.dayOfYear } }
 }
